Added tests for resolution flag (#2197)
* Added tests for resolution flag

Added test cases to make sure the resolution flag does what its supposed
to. It queues up a handful of common 16:9 and a couple 4:3 resolutions
and tries to render all of them with all acceptable separators. Once
they render it checks to make sure that the rendered resolution matches
the requested resolution.

* Speed up resolution tests

Simplified the scene rendered for the resolution flag tests to speed up
the tests. Switched over to `NoAnimation` scene and decreased the length
of the `NoAnimation` scene from 1 to 0.1.

* Convert paths to strings in test_resolution

For older versions of python on windows it seems having a `Path` object
in the list of command parameters causes issues. This patch explicitly
converts the path to a string for this case.

* Convert paths to string in  `get_video_metadata`

For older versions of python on windows it seems having a `Path` object
in the list of command parameters causes issues. This patch explicitly
converts the path to a string inside the `get_video_metadata` method.

Co-authored-by: Benjamin Hackl <devel@benjamin-hackl.at>