Allowed ``networkx`` to return 3D layouts when passing ``dim=3`` in the ``layout_config`` of a :class:`.Graph` (#2927)
* fix: return 3D points when layout_config has dims=3

When layout_config option dims was set to 3, changing or setting graph layout (via _determine_graph_layout) now returns a dictionary of tuples of length 3 to represent vertices. Previously returned a layout with tuples of length 4 instead. When the layout option dim is not 3 or unset, no change is made.

* Update manim/mobject/graph.py

Co-authored-by: Benjamin Hackl <devel@benjamin-hackl.at>