fix: Changed default CLI arguments to None
Config file settings were not being respected for `renderer`,
`force_window`, and `preview_command` CLI arguments.

This was because default values were defined in their cloup option
definitions, which overwrote the previously parsed config values from
any files in the `config` object of the `render` function.

This commit changes the default values for these arguments to `None`
so that the `config` object does not overwrite its existing values
with cloup defaults.